Implement parameter definition in a central place
Parameter definitions are defined in a separate header file which gets
included based on a compiler macro. A default header file with the
(ATLAS) parameter definition is provided.
Basic consistency checks for the supplied parameter definitions are
included (e.g. checking for the existence of certain typedef's and local
parameter constants).