not all abort traits need a result type
The type_collector helper should gracefully accept types which do not
provided the nested typedef required by the type extractor. The
motivation for this is that this behaviour avoid the necessity to
declare empty structs as result_type in observers which do not need a
result type (e.g. just debug output).