setting parameter values requires a non-const reference
For example, SingleCurvilinearTrackParameters::set uses
SingleTrackParameters::getParameterSet to update parameter values and
therefore requries a non-const version. For the moment, we make this
method protected to not expose this functionality to the user. The
reason being that it is quite simply to get the track parameters object
into an inconsistent state.