make dependencies of classes more clear
The propagation direction enum and the result struct does not depend on
the actual type of the propagation stepper. Therefore, there is no
reason to make them nested enums/structs.

The result object is now completely default constructible and using
unique pointers for storing the final propagation result. This allows to
initliaze them only at the end of the propagation.