guard against numerical jacobian evaluation against invalid values
When calculating the jacobian matrix numerically, we must make sure that
the theta parameter is inside a valid range after applying the epsilon
variation. Setting the theta parameter for a track parameter will cut it
off at the boundaries 0 and pi if it exceed those, leading to an smaller
effective step size. However, when calculating the numerical derivative,
the original step size is assumed.