update to boost-1.62 to fix warnings with gcc-6.2
boost-1.61 generated a warning about a faulty placement-new which is
fixed in boost-1.62. Unfortunately, cmake <3.7 does not handle
boost-1.62 properly and creates a warning during configuration. This
warning appears to be harmless. It is fixed in cmake >=3.7. This
(very recent) cmake version is not yet available on many OS. Given
that the warning has no severe consequences, we refrain from updating
the minimum cmake version.