Removed mutable from SingleCurvilinearTrackParameters
Replaced thread-unsafe lazy initialization of m_upSurface with eager
initialization on class construction.

Note that this version will only produce the same results as the
previous code if the track parameters' position and momentum are not
allowed to change. Otherwise, m_upSurface would need to be updated
accordingly, perhaps through some Observer pattern.

The reasons why I think this modification is safe are that the problem
already existed in SingleBoundTrackParameters, which does not employ
lazy initialization, and that there are currently no methods which allow
one to modify the position and momentum of TrackParametersBase and
SingleTrackParameters.