Removed mutables from TGeoBuilder (as much as current design will allow)
TGeoLayerBuilder has an ownership problem: it returns, by design,
pointers to members of a heap-allocated struct.

This means that it needs to become the owner of said struct, which is
both thread-unsafe (an internal vector-of-pointers must be updated on
every read) and a memory leak (no one will be able to discard the struct
once the returned pointer is discarded).

This problem is known, abundantly documented with code @todos, and will
hopefully be fixed eventually. But until then, the best I can do is to
restrict the scope of the internal mutable struct with a const_cast.