print results/attempts for finding dependencies with cmake
Most dependencies are optional. In case they can not be found, the
dependent parts of ACTS are disabled and a warning message is
printed. To keep the configuration output short and comprehensive
the cmake output from the attempt to find dependencies was silenced.
However, it sometimes is desirably to install a certain dependency.
In this case, the cmake output is actually very helpful to determine
why a certain dependency could not be fulfilled. Thus, cmake output
from finding external packages is enabled now by default. This
behavior can be overwritten by specifying

-DACTS_FIND_QUIET=ON

during the cmake command.