Homogeneized the side-effects of convert()
Here, the discrepancy was not between the AtlasStepper and the
EigenStepper, but between the two overloads of convert() for the
EigenStepper, which had different side effects.

One version modified the jacobian in-place, while the other performed a
computation on a copy of the jacobian. I believe that the latter
solution is the correct one, because if convert() trashes the value of
the Jacobian, the alterations performed by step() on said Jacobian
become mathematically invalid.