change BinningData lookup value wrap
for closed binnings: if value outside of boundary on one side, the value is wrapped. this replaces the logic were ‘out of bounds’ values were assigned to the outermost bin on the other side of the range.
also adds some unit test checks which explicitly assert this (for the phi case at least)

for open binnings, behaviour remains as is.