changes the way input is provided to LC and SAC
- introduce new struct ProtoLayer which holds min/max values
  - can calculate them in one loop, or have them manually set
- LC creates them, SAC accepts them but can also create them on the fly
- SAC respects the proto values for the binning
- LC can additionally apply envelopes for Layer dimensions

The calculation of extents is either done in ProtoLayer, or set in the caller of LC. Bin positions are always set in SAC, making the placement of the binning boundaries for phi closed consistent.

SurfaceMatcher is also made a config property in a newly introduced SAC config struct, cleaning up the function signatures.

Updates SAC unit tests