Add LCG93 and LCG94 builds an unit tests
Detect c++17 compatible compiler and compile with c++17

ROOT under a compiler with c++17 support and a flag set bakes in
\#defines which trigger header code to use c++17 features. That means
ROOT compiled with a c++17 compatible compiler results in code that can
only be compiled with c++17. CMake will now detect this, and set the std
to 17 *EVEN THOUGH* we do not use any c++17 features (yet).
Compatibility should be fine.

Refuses to compile on LCG93/CC7 with TGeoPlugin enabled,
since there is now good way to set the C++17 standard:
CXX_STANDARD=17 is not supported by CMake<=3.7.0
If possible, set c++17 as a compile feature on TGeo so it propagates
to dependees.
Disable TGeo plugin for cc7_lcg93 build