Implemented a git-hook-able clang-format check
This script is intended to be used as a local git commit hook (or can be
called by the user).
It checks all staged files with the clang-format in the docker
container and fixes any files if needed.
The files are then readded to the staging area.

It does not (!!) work with partially staged files.