phi, perp, theta, eta: assert dimensions
If static eigen types are given, a static asserts on the required number of
dimensionts is performed. No runtime check needs to be done here In case of
dynamically sized types, a runtime check is performed, and if it fails, program
execution is aborted (as opposed to silently returning 0. before)