Add OBB to Volume
The OBB is always in the "natural" frame of the volume, so it's invariant wr.t.
to the volume transform. The getter method returns a reference to the same
instance every time. To put yourself into the OBB system, you need the
inverted transform of the volume. This is now cached and stored on the volume,
and there is a getter.