Unify abort condition signature check and call pattern
The detection mechanism appears to have been not quite in sync with the
call mechanism. This rewrites the detection mechanism to be simpler and
correct:
- IF the aborter does not have a member type `action_type`, the check requires
  the operator to take `const propagator_state_t&, const stepper_t&`
- IF the aborter has a member type `action_type`, the check requires
  the operator to take
  `const propagator_state_t&, const stepper_t&, const result_t&`

The result argument is now last, which is much more intuitive. An
aborter can declare both overloads, but only the checked one will ever
be called by the implementation.

This also adds explicit tests of this behaviour.