Make the material classification numbers an opaque vector (#299)
Remove the public `enum` within the `Material` class that describes the content of the classification numbers vector. The classification numbers can still be used to stored the encoded material information. In order to interpret it, a `Material` object has to be constructed again.

The functionality was no used in the code base and removing it will make it easier to change the internal parametrization as planned in #31. Should also fix compiler warning regarding enum name clashes that show up in #267.