Fix installed CMake config (#267)
This fixes the CMake config files that are provided when Acts is installed. Components are now correctly listed and dependencies are actually searched for. A bare-bones downstream project is added that demonstrates how to use Acts as an external package. This project is build as part of the CI to test the installed config.

Include the following additional fixes:
- Eigen is now expected to provide a CMake config file. The manual `FindEigen` module has been removed.
- Fatras include files are installed correctly in `include/ActsFatras` instead of `include/ActsFatras/ActsFatras`.
- Finding packages is explicitely done in Config mode wherever possible.
- Correct in-source documentation for Acts-specific CMake modules.
- Use recommended the CMake config installation location and adopt names consistent with the CMake documentation.

Fixes #27. Fixes #265.