Stepper update methods without BoundParameters (#297)
Currently the (C)KF needs to build multiple `BoundParameters` objects in order to update the parameters and the covariance matrix in the stepper state. This PR allows implements an explicit update method for this case. The old method becomes therewith deprecated. The corresponding methods in the `MultiTrajectory` become functions that deliver the parameters in global representation in order to be passed to the stepper state. Alongside a reinitialisation of the Jacobian in the `AtlasStepper` is removed in order to synchronise the behaviour of the `update()` method with the other steppers.

Co-authored-by: Moritz Kiehn <msmk@cern.ch>