Introduce volume oriented Measurements (#272)
This PR introduces the concept of a `Measurement` that is located in a volume. Compared to the current `Measurement` object, that holds a `std::shared_ptr<const Surface>` and bound parameters, it gets extended to also storing a `Volume` pointer and free parameters. The underlying concept follows thereby the idea from `ParameterSet` and allows the deduction dependent on the used parameters indices. Along with this implementation an equality operator for `Volume` objects is required and part of this PR.
This PR only allows the construction of the new `Measurement` objects. These new types are not applied anywhere. This PR is a part of #165.