Release workflow update (#318)
- Remove a bunch of old stuff and scripts we don't need anymore
- New updated release helper script in `CI/release.py`
  - Can make a release branch, bump version and push, both for fix and minor releases.
  - Can make release notes from milestone and tag name (I've been using this for the releases so far on Github)
- Workflow file with a manual trigger to run exactly this. Once this is merged into master, we can manually trigger this workflow with a desired tag name. The script will figure out what to do (either create a new release branch, bump and tag there, or use existing release branch), make a release, create release notes and close the milestone.

I tried to make it assert as much as possible to prevent any data corruption, but the entire process is not a transaction. If for whatever reason it fails midway, manual cleanup is required. Anyway I think this might help automate this more. I tested this on my fork, and in the tests it seemed to work fine.

I think anyone with write access will be able to trigger the workflow, and thus make a new release.

Co-authored-by: robertlangenberg <56069170+robertlangenberg@users.noreply.github.com>