Moved over all the previous developments from cuda-improvements-20200612
Replaced the CUDA memory handling types with sleeker ones, streamlined
the code around the new types a little, and tweaked the CMake configuration
a little bit.

Renamed the CUDA seed-finder class to Acts::Cuda::Seedfinder. As there
was really just no reason for this type to be a template specialisation
of the Acts::Seedfinder type.