Hid the implementatio of the matrix and vector types from the outside.
In order to make it possible to provide a pure C++ interface for these
types, I also introduced the Acts::Cuda::StreamWrapper type. It will
be used to manage cudaStream_t objects in memory. But until it is
adopted explicitly all around the code, it does not delete the stream
objects just yet.