Fix volume material being double mapped on surface and volume (#369)
Fix an issue that would lead to volume material being mapped both on surface and volume in some edge case.
To avoid this form happening a check on whether the material was or not in volume was in place but the volume being considered only changed if at least one material hit was inside. This resulted in double counting if no material hit where in the first material volume crossed (see example plot from the ATLAS detector below).
To solve this issue we now change the volume being considered when we are not inside and further away from the origin than the entry point in said volume.