Added the ability to look for, and filter out duplicate spacepoints.
It seems that some input files are not perfect, they contain multiple
spacepoints with the exact same spacial coordinates. This makes the
seed finging's reproducibility on a GPU pretty much impossible.

At the same time added a print operator for TestSpacePoint, as I was
using it for debugging while writing the duplet searching code.