Add a "second" CUDA based SeedFinder implementation, master branch (2020.08.11.) (#371)
* Made the CUDA CMake configuration a little nicer.

Specifying explicitly which architectures to generate code for,
specifying that debug builds should attach debug symbols to both
the host and device CUDA code, and making all of these configurations
overridable from the command line.

* Introduced the ActsCuda2Plugin library.

In order not to interfere with the previously implemented
CUDA based seed finding, I've put my own implementation into
a separate plugin library.

The separate library is necessary because this library uses
a simpler build setup than ActsCudaPlugin.

* Added the Acts::Cuda::Details::countDublets function/kernel.

This code is used to figure out the exact parallelisation for the
triplet finding code.

* Moved Acts::SeedFilterConfig into its own header.

To make it easier to use it in the CUDA plugin code.

* Added the triplet finding code for the "new" CUDA seed finding.

* Now setting CMAKE_CUDA_FLAGS_DEBUG as a regular variable instead.

I forgot that the variable would be pre-defined as a cache variable
by CMake itself. So my setting was not doing anything.

* Introduced Acts::Cuda::SeedFinder.

As an alternative to Acts::Seedfinder<..., Acts::Cuda>. Unfortunately
the code still behaves a bit worryingly, but at least it more or less
works...

* Introduced a mechanism for providing seed filtering from the user's code.

The setup is a little clumsy, as function pointers are a bit
difficult with CUDA, but this seemed to be the most reasonable setup.

* Introduced a != operator for Acts::BinnedSPGroupIterator.

* Introduced a "full test" for Acts::Cuda::SeedFinder.

Based very much on the existing seedfinder tests. Only modifying
it to make it a bit sleeker in some places, and to accommodate the
setup requirements of Acts::Cuda::SeedFinder.

* Switched back to useing hardcoded filtering rules.

For some reason function pointers do not work correctly with
optimised CUDA builds. (I developed the code on a debug build...)
The results become incorrect without any error from CUDA, in such
optimised builds. :-/

* Slightly reduced the memory requirements of the triplet finding.

I forgot about improving the code this way after introducing the
Acts::Cuda::Details::DubletCounts::maxTriplets variable.

* Disabled the checks for the function pointers.

Forgot to remove those assert statements when I removed the usage of
the user defined filter functions. (As these do not work correctly
in optimised builds for some reason. I'm following up about that
at the moment...)

* Added the ability to look for, and filter out duplicate spacepoints.

It seems that some input files are not perfect, they contain multiple
spacepoints with the exact same spacial coordinates. This makes the
seed finging's reproducibility on a GPU pretty much impossible.

At the same time added a print operator for TestSpacePoint, as I was
using it for debugging while writing the duplet searching code.

* Turned ActsCuda2Plugin into a static library.

This is necessary to make it possible to give used-defined filter
functions to the triplet finding code. (It's a long-standing
issue in CUDA that device code split across shared libraries does
not work well...)

* Fixed the duplicate spacepoint removal in ActsUnitTestSeedfinderCuda2.

Turns out that the code was only counting the number of spacepoint
duplicates so far, it was not discarding them...

At the same time updated the name of the executable to be in line with
Andi's "new naming".

Co-authored-by: robertlangenberg <56069170+robertlangenberg@users.noreply.github.com>