Material mapping onto CutoutCylinderVolume (#389)
Added the possibility to adjust the binutility to CutoutCylinderVolume this will be necessary to mapped material onto the ATLAS calorimeter for example.

Also updated the binutility created by default in the json converter. All the phi bin use to be closed, now they are closed if eHalfPhiSector is equal to Pi.