Made the CUDA seed finding process triplets for multiple middle spacepoints at once.
This was done with a **lot** of different changes, which were developed in
a separate branch. This is just a cleaned up version of all of those
developments.

The code now includes the ability to use CUDA streams, and now manages
CUDA device memory using its own manager class (Acts::Cuda::MemoryManager).