build!: ensure consistent CMake options and target names (#406)
Rename plugin-related build options to `CMAKE_BUILD_PLUGIN_...` and the corresponding targets to `ActsPlugin...` to be consistent with the rest of the options e.g. `CMAKE_EXAMPLES_...` options and `ActsUnitTest...` targets.

The option to search for an external `nlohmann::json` library is inverted and renamed to `ACTS_USE_SYSTEM_NLOHMANN_JSON`. With this change, all options are be off/empty by default without any special cases.

BREAKING CHANGE: All option and target names for the plugins are changed, so build instructions/scripts must be updated and targets linking against any of the plugins need to be changed to link against the new names.