feat: extend track parameters accessors (#428)
This adds the following additional accessors for derived quantities to all track parameters:

```cpp
.fourPosition(...)
.unitDirection()
.transverseMomentum()
.absoluteMomentum()
```

Having separate access to the direction and the absolute momentum provides the quantities that are actually used e.g. by the propagators. It is also closer to what is stored in the parameters vectors and reduces unnecessary computations.

This also removes the zero-arg `.position()` access for the `SingleCurvilinearTrackParameters` since it can not be consistently implemented without requiring a default geometry context.

The parameters concepts are updated to require `.fourPosition(...)`, `.unitDirection()` and `.absoluteMomentum()` instead of `.momentum()`. All track parameters retain the `.momentum()` accessor as a convenience for users, but it is not used internally in the code base.