fix: Calorimeter navigation bug (#487)
Fix the error that occur in the extrapolation trough the ATLAS calorimeter. If possible I would like to have this in before we update Acts in the ATLAS external.

The issue is the following :  During the extrapolation the size of the step sometime reach 1e-13 which lead to a too many step error. Looking at the log what seem to appear is that the next surface and the current surface are the same, the step size thus become 0 and the navigation get stuck. 

In the navigation code there is a condition to avoid this situation : state.navigation.currentSurface != protoNavSurfaces.front().object, this work because the surface are ordered by path length but if two surface are at the same position the ordering might not be correct. I thus instead we request the distance to the previous surface to be larger than 1um.