refactor(examples): Smear particle in the perigee frame (#496)
Change the `ParticleSmearing` to actually smear in the perigee frame with a common perigee surface for all particles originating from the same vertex instead of smearing each particle in its own curvilinear frame with projected uncertainties.

To do so, the default track parameters in the framework are changed to bound parameters. With the recent changes of the Core EDM this is no possible since curvilinear parameters are just bound parameters and can be store within.