feat: Add an Autodiff Plugin (#384)
This PR adds a Plugin, which allows to do the transport of the covariance matrix with methods of automatic differentiation instead of computing the derivatives by hand. Essentially, it provides the `AutodiffWrapperExtension` which can be used with a templated extension in the `EigenStepper`. The necessary templated extensions can also be found in the plugin-folder.

The PR introduces:
- A new header-only plugin
- A new third-party dependency (autodiff)
- Integration tests for the AutodiffPropagator