fix: Fix navigation when collecting surface in tracking volume hierarchy (#501)
This PR solve an issue with the navigation trough the ATLAS calorimeter (or any other volume using `compatibleSurfacesFromHierarchy`). If the none of the surface collected between start and the target intersect with the trajectory the navigation cannot reach the Navigation break (and thus cannot reach the target surface) and will loop forever.

To solve this issue an extra bool was added to the navigation state, after iterating over all the surfaces if no layer exist (which mean the surface where obtain via `compatibleSurfacesFromHierarchy`) the bool is switched to true and the layer stage is skipped.