refactor!: Set source link calibrator via Kalman/Ckf options (EDM rework part 1/4) (#510)
Move the `calibrator_t` typedef from the Kalman/Ckf types to the respective option types. In general, the calibrator is specific for a given source link implementation and there is no well-defined default behavior. The source link is also only known at call time and might not be defined a construction time. Moving the calibrator to the options means it can be set to match the source link and include information that might only be known at runtime.

The previous default calibrator `VoidMeasurementCalibrator` relied on specific behavior of the `MinimalSourceLink` that is not part of the source link concept. To avoid confusion, the calibrator must now always be set explicitly via the options types and the default implementation has been removed. For use-cases based on the `MinimalSourceLink`, a `MinimalSourceLinkCalibrator` is provided instead.

Additional cleanup: `outlier_finder_t` for the `KalmanFitter` and `source_link_selector_t` for the `CombinatorialKalmanFilter` were previously defined both for the implementation and for the option types. To avoid possible inconsistencies they are now only set via the option types as well.

BREAKING CHANGE: The `KalmanFitter` loses its `calibrator_t` and `outlier_finder_t` template types. The `CombinatorialKalmanFilter` loses its `calibrator_t` and `source_link_selector_t` template types. In both cases, the types are now set via the respective options types and need to be known only at call time instead of already at construction time. The default `VoidMeasurementCalibrator` has been removed.