refactor!: Move MinimalSourceLink to tests (EDM rework 3b/4) (#533)
BREAKING CHANGE: The `MinimalSourceLink` is removed from the public interface.

Users are expected to provide their own concrete source link implementation. The provided `MinimalSourceLink` is a very rigid example implementation that only allows a narrow use case. It is intended to be used for tests and only used as such in the code base. The examples framework uses a different and completely independent source link type. To avoid confusion for users and clarify its purpose, the `MinimalSourceLink` is removed from the public interface, renamed to `TestSourceLink`, and moved to the common test helpers.