feat(Examples): add csv simhit writer/reader (#537)
Add a dedicated simhit csv reader/writer pair in the examples, so as to be able to write out simhits before digitization and independent of clusters. In conjunction with that, add a simhit data structure in the TrackMlData to avoid setting the hit id, which becomes available only after digitization. The simhit writer was integrated in the Fatras Simulation Example for testing.