feat: Geant4 event recording (#516)
This PR provides the possibility to produce an entire record of (most) key parameters in a Geant4 event simulation. The event is thereby translated into the HepMC3 data format. The corresponding writer of HepMC3 events has been adapted within this PR to allow it's application in the `Sequencer`. Due to symmetry reasons, the corresponding reader has been extended by the same feature.