feat: Digitization 1, two-dimensional channelizer  (#542)
This PR is the first of a series to prepare for the new digitisation schema.

It introduces a two-dimensional channelized (in contrast to the 3D stepper used so far), that will allow digitisation in pixels, strips in both cartesian and polar coordinates. This PR only contains the channelized, i.e. a(n already drifted and clipped) surface segment is divided into its channel segments for further channel merging and/or clusterization.