refactor!: Simplify interface for jacobian between local and global parameters (#534)
BREAKING CHANGE: This PR simplifies the calculation within the covarianceTransport by changing the surfaceDerivative method to localToLocalJacobian which packs up the calculation for the full jacobian from local parameters at start surface to local parameters at the final surface. Among the localToLocalJacobian input parameters, only the jacFull is updated. Reinitializations of other jacobians are done separately within the covarianceTransport. In this way, how different jacobians are updated with each covarianceTransport is more clear.

This PR also packs the global position and direction to a FreeVector as the input parameter of initJacobianToLocal and initJacobianToGlobal. This could make the interfaces for jacobian/derivative calculation more consistent.