perf: allow reco for DD4hep detector (#549)
Currently, the track reco is using sim hits read in by the `CsvPlanarClusterReader`, but the `PlanarSteppingAlgorithm` is only working for the `GenericDetector`. Hence we only have clusters info for the `GenericDetector`, which blocks the track reco for other detectors.

This PR replaces the `CsvPlanarClusterReader` with `CsvSimHitReader` to read in directly the sim hits (written out by the `CsvSimHitWriter` after simulation) for the track reco. The sim hits will be passed to hit smearing for further processing.

It also implements a common interface for track reco with different detectors and add optionally the reco executables for DD4hep detector.