fix: AMVF track linearization and seed constraint fix for q431 test (backport) (#565)
In case the algorithm finds a bad vertex and deletes it at some point again, I need to force it to relinearize the tracks in the next iteration since they were linearized with respect to the bad vertex. If the seed constraint is turned off, the old (e.g. beamspot) constraint will now be correctly applied. For consistency reasons with Athena, a negative determinant check in the track linearizer is added.