refactor: Clean up ImpactPointEstimator tests (#567)
This is a clean up/ rewrite of the `ImpactPointEstimator` tests in order to understand and fix the problem discussed in #455. The iteration over multiple tracks is now handled via data-driven tests, common code has been moved into helper functions, and the multi track compatibility test has been removed.

The issue discussed in #455 happens in the multi track compatibility test: The vertex distance and compatibility score (chi2) for multiple (random) tracks is calculated. For all possible track combinations the (relative) difference in real distance and in compatibility score was computed. The test then checked that both differences have the same sign. This is equivalent to testing that a track w/ a smaller real distance to the vertex also has a smaller score. This was the test that failed and was previously disabled.

After further investigations, it became clear that this test does not make sense to begin with and that the apparent success was due to a lucky choice of random numbers that lead to generated tracks that just happened to satisfy this. Changing the random number seed and/or changing the number of generated tracks randomly changes the point where the error occurs.

The score is computed as the 2d-distance in the curvilinear plane at the closest point weighted by the propagated track covariance. Thus, a track with a small real distance but a tiny covariance would still get a bad score, while a track with a large real distance but a huge uncertainty would get a good one. This is true even for the tested case where the initial covariance in the perigee frame is identical for all tracks. Due to the random orientation of the track, the projection into the plane used for the score computations does not need to be the same.

Closes #455.