fix: AMVF track linearization and seed constraint fix for q431 test (master) (#550)
This PR targets master and fixes the remaining issue we see in the q431 job in the CI of this Athena MR where in one single event 17 jets (ACTS AMVF enabled) are reconstructed instead of 15 jets (Athena AMVF enabled):

In case the algorithm finds a bad vertex and deletes it at some point again, I need to force it to relinearize the tracks in the next iteration since they were linearized with respect to the bad vertex.
If the seed constraint is turned off, the old (e.g. beamspot) constraint will now be correctly applied.
For consistency reasons with Athena, a negative determinant check in the track linearizer is added.
The corresponding backport PR for a bug fix release v.1.2.1 can be found here.