refactor: HepMC event record reduction and selection (#541)
The Geant4 event records can become very large. In order to reduce the memory consumption a filter is added that allows the combination of multiple steps with certain processes (e.g. Transportation) into a single vertex. Furthermore this PR adds a list to select or reject events with certain processes.