refactor: Allow running (C)KF in outward->inward mode (#551)
Currently, the (C)KF assumes the Kalman filtering is done in inward->outward mode with the seed around the IP. And It always use sthe last smoothed track parameter to the target surface.

This PR refactors the (C)KF to make it also working for the outside->inside mode with a seed at the tracker exit. It also uses the fitted track parameter adaptively (not always the last smoothed one) to propagate to the target surface.