fix: Direct Navigation KF (#521)
This addresses the bug raised in Issue [508](https://github.com/acts-project/acts/issues/508). It is a a continuation of a closed [PR](https://github.com/acts-project/acts/pull/512). The directed navigation fit function in the KF had not been updated with some of the recent refactoring changes. This fell through the cracks since there is not a specific example algorithm that runs the directed navigation KF fit function.

This PR modifies the `FittingAlgorithm` to run either the standard Navigation or the directed Navigation and adds a command line option in the `RecTruthTracks` executable to choose which kind of navigation should be run in the fitting.