refactor: Refactor KalmanFitter and CKF tests (#585)
The measurement creation code is split into a common helper. The `KalmanFitter` test is split into multiple test cases where each one tests only a single configuration. The results for both `KalmanFilter` and `CKF` tests are systematically verified.

This PR is a preparation for the EDM rework in #557 by reducing the number of places, where the implementation details of the `TestSourceLink` (and thus the underlying measurement) need to be considered.