fix: Fix linearization point updates in AMVFitter (#586)
This PR fixes a last issue that was seen when running on data events in Athena, where the linearization point needs only to be updated in certain cases (i.e. not if the track is not linearized).
Also a small track weight cut is applied for the vertex smoothing.
Both changes have only a very small impact on the AMVF outcome, but are still needed for it to function properly.