ci: fix macosx build (#611)
Set `CMAKE_CXX_STANDARD=17` explicitly for the macOS CI build.

The DD4hep CMake configurations seems to be buggy (surprise). It seems to require that the `CMAKE_CXX_STANDARD` options is set to a definite value, but this only gets triggered on recent CMake versions. The current CMake version installed via homebrew seems to do just that.