refactor!: Remove/replace ParameterSet-based measurement (EDM rework 4d/4) (#557)
BREAKING CHANGE: Measurement is replaced with a new implementation that is templated only on the measurement dimension and does not store a reference object anymore.Fittable{Volume}Measurement is replaced by new variants {Bound,Free}VariantMeasurement based on the updated Measurement type. ParameterSet and related types are removed. Since the Measurement does not store the reference object anymore, setting the calibrated measurement for a MultiTrajectory state does not automatically set the reference surface anymore. The reference surface must always be set explicitly.

In addition to the core changes, this includes quite a few dependent changes in Fatras, Examples, and Tests to adapt to the new measurement types:

Rewrite the Fatras smearing digitizers. The are now also templated only the the measurement size with the measurement subspace configurable at runtime.
Rewrite the smearing configuration in the Examples to match the updated Fatras types.
Examples algorithms that handle measurements and require access to the local-to-global transforms now take require a TrackingGeometry pointer via their configuration structs. It is required since the Surface is not available anymore directly via the measurements.
The TestSourceLink in the unit tests now stores the measurement data inline. Previously, the TestSourceLink just stored a pointer to a measurement which results in a circular dependency, i.e. TestSourceLink points to FittableMeasurement which stores a copy of TestSourceLink. This worked before by storing fake information in at least one of the objects involved. It does not work anymore with the updated Measurement type. The updated TestSourceLink also matches the intended data flow for source links, i.e. you have the source link first and then use a calibrator to retrieve the Measurement.
Contains #554 and #555 and must be merged afterwards. Finally closes #350.