refactor(Examples): Introduce multi-value option types with consistent encoding (#633)
Switch all multi-value options in the examples to use the x:y:z encoding scheme. This is achieved by introducing custom types with manual en-/decoding to/from strings. Negative values require no special handling anymore and the types can be used to put further restrictions on the type and number of parameters.