refactor(Examples): Remove Hep{PDT/PID} dependency (#638)
Remove the dependency of Examples/Io/HepMC3 on the HepPDT package. HepPDT is obsolete and not supported. It was used to lookup particle charge and name for a PDG id in the particle data tables. That functionality is already available in ActsFatras without external dependencies and is used instead.

HepMC-related code in the examples are now part of the everything-build in the CI since there are now compatible HepMC versions available on all CI jobs.

Includes some fixes in the HepMC3 module to make it compile in the CI and some drive-by whitespace fixes for consistent formatting.

Closes #456.