build: Update dependency versions and build documentation (#639)
All dependency versions are bumped to the minimal versions available in the CI builds. They are the only versions where we can ensure that they work. For the core, the minimal CMake version is bumped to 3.14, Eigen to 3.3.7, and Boost to 1.71. This enables the removal of some CMake work-arounds that were previously necessary.

Additional changes:

-   The setup scripts for lxplus/CVMFS are updated to match the CI LCG releases. Scripts for all other versions are removed to avoid confusion.
-   Cleanup and documentation updates in various CMake-related files, e.g. optional components and subdirectories are consistently logged only if they are activated.
-   The documentation is updated to match the actual requirements, available containers, etc.

This closes #635.