refactor!: optimize CKF (#644)
This PR includes the following changes:

- Change the CKF `findTracks` interface:
The function now takes all the seeds as one of the input parameter instead of just one seed. In this way, all the seeds could use the same `inputMeasurements`, i.e. we don't have to fill the `inputMeasurements` for each seed, which is quite time-consuming. In this way, the CKF timing performance could be improved.

- Change the `CkFSourceLinkSelector` to `MeasurementSelector`:
The current `CKFSourceLinkSelector` takes all the source links on the surface as one input and calibrate them before the selection. The calibration is then done again in the CKF for the selected source links. To remove the duplication of the calibration, the selector now takes the calibrated measurement as inputs.

- Fix the CKF unit test in backward propagation mode