feat: introduce Json converter for SurfaceBounds (#660)
This PR includes the next Json converter, this time for SurfaceBounds.

All SurfaceBounds have a values() streaming method (providing a std::vector) and a constructor by providing a std::array. Hence, only one converter is needed, however, the type has to be given explicitely, no automated conversion is possible, as the Surface objects never swallow SurfaceBounds directly but a class extension of the appropriate type.

UnitTest for one is attached, no need to call the same functions on all inherited objects.